backend/pickup-points/requirements.txt
psycopg2-binary==2.9.9
qrcode==7.4.2
Pillow==10.2.0